+++ database/migrations/2021_05_11_033231_create_anexos_table.php
@@ -17,7 +17,7 @@ class CreateAnexosTable extends Migration
             $table->increments('id');
             $table->integer('tipo');
             $table->text('arquivo');
-            $table->integer('codigo_chamado');
+            $table->unsignedBigInteger('codigo_chamado');
             $table->timestamps();
 
             $table->foreign('codigo_chamado')
