+++ database/migrations/2021_05_11_033242_create_problemas_table.php
@@ -16,7 +16,7 @@ class CreateProblemasTable extends Migration
         Schema::create('problemas', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('codigo_categoria');
+            $table->unsignedBigInteger('codigo_categoria');
             $table->timestamps();
 
             $table->foreign('codigo_categoria')
